+++ src/data/user-monitoring/permission-fixer/D2ExternalConfigRepository.ts
