+++ src/domain/config/repositories/UserMonitoringConfigRepository.ts
@@ -1,4 +1,4 @@
-import { UsersOptions } from "domain/entities/user-monitoring/UserMonitoring";
+import { UsersOptions } from "domain/entities/user-monitoring/common/UserOptions";
 
 export interface UserMonitoringConfigRepository {
     get(): Promise<UsersOptions>;
