+++ src/domain/entities/user-monitoring/common/AuthOptions.ts
@@ -1,0 +1,3 @@
+export interface AuthOptions {
+    apiurl: string;
+}
