+++ src/domain/entities/user-monitoring/common/Identifier.ts
@@ -1,0 +1,8 @@
+export interface Item {
+    id: string;
+    name: string;
+}
+export type Id = string;
+export type IdItem = {
+    id: Id;
+};
