+++ src/domain/entities/user-monitoring/common/RolesByGroup.ts
@@ -1,0 +1,6 @@
+import { Item } from "./Identifier";
+
+export interface RolesByGroup {
+    role: Item;
+    group: Item;
+}
