+++ src/domain/entities/user-monitoring/common/RolesByRoles.ts
@@ -1,0 +1,6 @@
+import { Item } from "./Identifier";
+
+export interface RolesByRoles {
+    active_role: Item;
+    ignore_role: Item;
+}
