+++ src/domain/entities/user-monitoring/common/RolesByUser.ts
@@ -1,0 +1,6 @@
+import { Item } from "./Identifier";
+
+export interface RolesByUser {
+    role: Item;
+    user: Item;
+}
