+++ src/domain/entities/user-monitoring/common/UserCredentials.ts
@@ -1,0 +1,16 @@
+import { Id, IdItem } from "./Identifier";
+
+export interface UserCredentials {
+    lastUpdated: string;
+    lastLogin: string;
+    passwordLastUpdated: StringDateTime;
+    invitation: boolean;
+    selfRegisterd: boolean;
+    uid: Id;
+    disabled: boolean;
+    twoFA: boolean;
+    username: string;
+    userRoles: IdItem[];
+}
+
+type StringDateTime = string;
