+++ src/domain/entities/user-monitoring/common/UserDetails.ts
@@ -1,0 +1,8 @@
+import { Id } from "./Identifier";
+
+export interface UserDetails {
+    id: Id;
+    displayName: string;
+    name: string;
+    username: string;
+}
