+++ src/domain/entities/user-monitoring/common/UserGroup.ts
