+++ src/domain/entities/user-monitoring/common/UserRoleAuthority.ts
@@ -1,0 +1,7 @@
+import { Id } from "./Identifier";
+
+export interface UserRoleAuthority {
+    id: Id;
+    authorities: string[];
+    name: string;
+}
