+++ src/domain/entities/user-monitoring/common/UserWithoutTwoFactor.ts
@@ -1,0 +1,7 @@
+import { Item } from "./Identifier";
+
+export type UserWithoutTwoFactor = {
+    invalidUsersCount: number;
+    listOfAffectedUsers: Item[];
+    response: string;
+};
