+++ src/domain/repositories/user-monitoring/common/MetadataRepository.ts
@@ -1,8 +1,6 @@
 import { ProgramMetadata } from "data/d2-users/D2Users.types";
 import { Async } from "domain/entities/Async";
-import { TemplateGroupWithAuthorities, UsersOptions } from "domain/entities/user-monitoring/UserMonitoring";
 
 export interface MetadataRepository {
-    getTemplateAuthorities(options: UsersOptions): Promise<Async<TemplateGroupWithAuthorities[]>>;
     getMetadata(programId: string): Promise<Async<ProgramMetadata>>;
 }
