+++ src/domain/usecases/user-monitoring/two-factor-monitoring/RunReportUsersWithout2FA.ts
@@ -1,10 +1,8 @@
 import { Async } from "domain/entities/Async";
-import {
-    UserMonitoringCountResponse,
-    UserWithoutTwoFactor,
-    UsersOptions,
-} from "domain/entities/user-monitoring/UserMonitoring";
-import { MetadataRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/MetadataRepository";
+import { UserMonitoringCountResponse } from "domain/entities/user-monitoring/common/UserMonitoring";
+import { UsersOptions } from "domain/entities/user-monitoring/common/UserOptions";
+import { UserWithoutTwoFactor } from "domain/entities/user-monitoring/common/UserWithoutTwoFactor";
+import { MetadataRepository } from "domain/repositories/user-monitoring/common/MetadataRepository";
 import { ReportRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/ReportRepository";
 import { UserRepository } from "domain/repositories/user-monitoring/two-factor-monitoring/UserRepository";
 import _ from "lodash";
